dodola-sample/requirements.txt
anyio==3.7.1
beautifulsoup4==4.12.2
certifi==2023.7.22
chevron==0.14.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
loguru==0.7.2
lxml==4.9.3
msgpack==1.0.7
multidict==6.0.4
nonebot-adapter-onebot==2.3.1
nonebot-plugin-gocqhttp==0.6.12
nonebot2==2.1.2
psutil==5.9.6
py-cpuinfo==9.0.0
pydantic==1.10.13
pygtrie==2.5.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sniffio==1.3.0
soupsieve==2.5
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.8.0
uvicorn==0.24.0.post1
watchfiles==0.21.0
websockets==12.0
win32-setctime==1.1.0
yarl==1.9.2
